requirements.txt
# Core dependencies
pydantic>=1.10
requests>=2.28

# CrewAI and LangChain multi-agent orchestration
crewai==0.141.0
langchain==0.3.20
langchain-community==0.3.19
langchain-openai==0.3.25

# Google Gemini integration
google-generativeai>=0.3.0

# CrewAI tools and utilities
crewai-tools==0.51.1
duckduckgo-search==7.5.2

# Configuration and utilities
ruamel.yaml>=0.17.0
pyyaml>=6.0
